pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "cass-logger-app"
version = "0.1.8"
description = "Desktop GUI application for the Cass data logger"
requires-python = ">=3.12"
authors = [
  { name = "Cameron Chaney" }
]
dependencies = [
  "cass-logger-dev @ git+https://github.com/mcharlesmorrison/cass_logger_dev",
  # cass-logger-dev runtime deps (parent pyproject.toml doesn't declare them)
  "pyserial>=3.5",
  "pandas>=2.2.3",
  "numpy>=2.1.3",
  "python-dateutil>=2.8.2",
  "fitdecode>=0.1.0",
  "DateTime>=5.5,<6.0",
  # GUI deps
  "pywebview>=5.0",
  "pyinstaller>=6.0",
  "platformdirs>=4.0",
  "certifi>=2024.0",
]

[project.urls]
Repository = "https://github.com/camchaney/cass_logger_app"

[tool.setuptools.packages.find]
where = ["."]
